remove get-config.js because only webpack can use fs
